Harmonized the ledger to u32 and timestamps to i64
